supabase/functions/paddleocr-service/requirements.txt
paddlepaddle==2.5.2
paddleocr==2.7.3
pdf2image==1.16.3
Pillow==10.0.1
numpy==1.24.3
opencv-python==4.8.1.78
